public_transport_app/requirement.txt
# ── Backend dependencies ─────────────────────────────────────
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.28.0
stripe>=5.0.0
python-dotenv>=1.0.0

# ── Mobile (Kivy) dependencies ──────────────────────────────
kivy>=2.1.0
kivy-garden.mapview>=0.7.4
plyer>=2.0.0
